Add files via upload
This is a script for my port scan automation to ease my daily port scanning on different targets, it is based on nmap so it will required nmap install. it uses -p- and -T4 for scanning for finding all the open ports and then uses -sV and -sC for depper scan on open ports only and then does -sU to not miss udp ports.
Please read the source code before using.